hierarchy restructure
