+++ lib/equipment_slots.py
