+++ maps/__pycache__/monster_dict.cpython-37.pyc
